refactor transformer, verify new fields are working
