+++ etl/transformer/vacant_table/vocabulary_data/commercial-building-construction-type.json
