+++ etl/transformer/vacant_table/vocabulary_data/residential-building-basement-type.json
