+++ etl/transformer/vacant_table/vocabulary_data/residential-building-exterior-wall-type.json
