+++ etl/transformer/vacant_table/vocabulary_data/residential-building-occupancy-type.json
